pyproject.toml
[tool.poetry]
name = "zmq_tubes"
version = "1.5.0"
license = "MIT"
readme = "README.md"
description = "Wrapper for ZMQ comunication."
authors = ["Martin Korbel <mkorbel@alps.cz>"]
homepage = "https://github.com/calcite/zmq_tubes"
repository = "https://github.com/calcite/zmq_tubes"
documentation = "https://github.com/calcite/zmq_tubes"

keywords = ["zmq", "mqtt", "tubes", "zmq_tubes"]
classifiers = [
    "Topic :: System :: Networking",
    "Topic :: Software Development :: Libraries :: Python Modules"
]

[tool.poetry.dependencies]
python = "^3.6"
pyzmq = "*"


[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
PyYAML = "^5.4.1 || ^6.0"
pytest-asyncio = [{ version = "^0.19.0", python = "^3.7" },
                  { version = "^0.16.0", python = "^3.6" }]


[tool.pytest.ini_options]
testpaths = [
    "tests"
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
